feat(router): Created router and proxy functions using Flask (py)
I have standardised the configuration for all the  services (see config.json), and then I went ahead and made a preliminary solution for the routing/proxy engine we will use. The idea here is that as people use our services, all the actual domain requests will be processed through this router. So, when there is a new user for example, the domain <username>.st.ax or whatever can easily be used straight away, no DNS modification is required. Whoop whoop �... pretty clever Alex.. :)